+++ backend/Makefile
@@ -36,7 +36,7 @@ index-data:
 	@echo "Indexing Nest data"
 	@CMD="poetry run python manage.py algolia_reindex" $(MAKE) exec-backend-command
 	@CMD="poetry run python manage.py algolia_update_synonyms" $(MAKE) exec-backend-command
-	@CMD="poetry run python manage.py add_suggestion_index" $(MAKE) exec-backend-command
+	@CMD="poetry run python manage.py algolia_update_suggestions" $(MAKE) exec-backend-command
 
 load-data:
 	@echo "Loading Nest data"
